Saving key state in Window
Improved key state implementation and sample

Improved handling of key events on Windows

Removed duplicated code in processEvents() and put it in a single
function.
Correctly handle Left and Right for special keys (Ctrl, Alt, System and
Shift)
Bool used to indicate if one of these special keys is pressed are now
set correctly.
The sample as been modified to show this.